Move flask code to separate module
